Implemented the search
The search functionality is implemented
It includes two searches.
In the first search the user will click on the only result. In the result they click on one of the key:value pairs which does an automatic search for data with that key:value pair.

They then do another search and show how you can use the 'copy link' functionality.

From here we could also show the sorting feature.